Redirect users to corresponding page of other version (not just overview)
I updated serve.py to build a map of pages (suffixes after the secion name) to their corresponding path. This is used to find matching pages in other versions of the doc. It doesn't work for all pages because there aren't always exact matches, but it does work most of the time.

This solves the frequent frustration of mine searching in Google, landing on v1.0 polymer docs, being told I'm out of date, clicking to change the version, then landing on an unrelated page, and needing to re-navigate to the page I wanted.